+++ fs/ceph/file.c
@@ -136,7 +136,8 @@ int ceph_open(struct inode *inode, struct file *file)
 		flags = O_DIRECTORY;  /* mds likes to know */
 
 	dout("open inode %p ino %llx.%llx file %p flags %d (%d)\n", inode,
-	     ceph_ino(inode), ceph_snap(inode), file, flags, file->f_flags);
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), file, flags, file->f_flags);
 	fmode = ceph_flags_to_mode(flags);
 	wanted = ceph_caps_for_mode(fmode);
 
@@ -201,7 +202,8 @@ int ceph_open(struct inode *inode, struct file *file)
 		err = ceph_init_file(inode, file, req->r_fmode);
 	ceph_mdsc_put_request(req);
 	dout("open result=%d on %llx.%llx\n", err,
-		ceph_ino(inode), ceph_snap(inode));
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode));
 out:
 	return err;
 }
@@ -637,9 +639,9 @@ static ssize_t ceph_aio_read(struct kiocb *iocb, const struct iovec *iov,
 	int want, got = 0;
 	int checkeof = 0, read = 0;
 
-	dout("aio_read %p %llx.%llx %llu~%u trying to get caps on %p\n",
-	     inode, ceph_ino(inode), ceph_snap(inode), pos,
-	     (unsigned) len, inode);
+	dout("aio_read %p %llx.%llx %llu~%u trying to get caps on %p\n", inode,
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), pos, (unsigned) len, inode);
 again:
 	__ceph_do_pending_vmtruncate(inode);
 	if (fi->fmode & CEPH_FILE_MODE_LAZY)
@@ -649,9 +651,9 @@ again:
 	ret = ceph_get_caps(ci, CEPH_CAP_FILE_RD, want, &got, -1);
 	if (ret < 0)
 		goto out;
-	dout("aio_read %p %llx.%llx %llu~%u got cap refs on %s\n",
-	     inode, ceph_ino(inode), ceph_snap(inode), pos,
-	     (unsigned)len, ceph_cap_string(got));
+	dout("aio_read %p %llx.%llx %llu~%u got cap refs on %s\n", inode,
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), pos, (unsigned)len, ceph_cap_string(got));
 
 	if ((got & (CEPH_CAP_FILE_CACHE|CEPH_CAP_FILE_LAZYIO)) == 0 ||
 	    (iocb->ki_filp->f_flags & O_DIRECT) ||
@@ -663,9 +665,9 @@ again:
 		ret = generic_file_aio_read(iocb, iov, nr_segs, pos);
 
 out:
-	dout("aio_read %p %llx.%llx dropping cap refs on %s = %d\n",
-	     inode, ceph_ino(inode), ceph_snap(inode),
-	     ceph_cap_string(got), (int) ret);
+	dout("aio_read %p %llx.%llx dropping cap refs on %s = %d\n", inode,
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), ceph_cap_string(got), (int) ret);
 	ceph_put_cap_refs(ci, got);
 
 	if (checkeof && ret >= 0) {
@@ -718,8 +720,8 @@ retry_snap:
 		return -ENOSPC;
 	__ceph_do_pending_vmtruncate(inode);
 	dout("aio_write %p %llx.%llx %llu~%u getting caps. i_size %llu\n",
-	     inode, ceph_ino(inode), ceph_snap(inode), pos,
-	     (unsigned) iov->iov_len, inode->i_size);
+		inode, (unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), pos, (unsigned) iov->iov_len, inode->i_size);
 	if (fi->fmode & CEPH_FILE_MODE_LAZY)
 		want = CEPH_CAP_FILE_BUFFER | CEPH_CAP_FILE_LAZYIO;
 	else
@@ -728,9 +730,10 @@ retry_snap:
 	if (ret < 0)
 		goto out_put;
 
-	dout("aio_write %p %llx.%llx %llu~%u  got cap refs on %s\n",
-	     inode, ceph_ino(inode), ceph_snap(inode), pos,
-	     (unsigned) iov->iov_len, ceph_cap_string(got));
+	dout("aio_write %p %llx.%llx %llu~%u  got cap refs on %s\n", inode,
+		(unsigned long long) ceph_ino(inode),
+		ceph_snap(inode), pos, (unsigned) iov->iov_len,
+		ceph_cap_string(got));
 
 	if ((got & (CEPH_CAP_FILE_BUFFER|CEPH_CAP_FILE_LAZYIO)) == 0 ||
 	    (iocb->ki_filp->f_flags & O_DIRECT) ||
@@ -775,15 +778,16 @@ retry_snap:
 
 out_put:
 	dout("aio_write %p %llx.%llx %llu~%u  dropping cap refs on %s\n",
-	     inode, ceph_ino(inode), ceph_snap(inode), pos,
-	     (unsigned) iov->iov_len, ceph_cap_string(got));
+	     inode, (unsigned long long) ceph_ino(inode),
+	     ceph_snap(inode), pos, (unsigned) iov->iov_len,
+	     ceph_cap_string(got));
 	ceph_put_cap_refs(ci, got);
 
 out:
 	if (ret == -EOLDSNAPC) {
 		dout("aio_write %p %llx.%llx %llu~%u got EOLDSNAPC, retrying\n",
-		     inode, ceph_ino(inode), ceph_snap(inode),
-		     pos, (unsigned) iov->iov_len);
+		     inode, (unsigned long long) ceph_ino(inode),
+		     ceph_snap(inode), pos, (unsigned) iov->iov_len);
 		goto retry_snap;
 	}
