+++ fs/ceph/super.c
@@ -896,7 +896,7 @@ static struct dentry *ceph_mount(struct file_system_type *fs_type,
 	if (IS_ERR(res))
 		goto out_splat;
 	dout("root %p inode %p ino %llx.%llx\n", res,
-	     res->d_inode, ceph_ino(res->d_inode),
+	     res->d_inode, (unsigned long long) ceph_ino(res->d_inode),
 	     ceph_snap(res->d_inode));
 
 	return res;
